+++ examples/request_redraw_threaded.rs
@@ -7,7 +7,7 @@ fn main() -> Result<(), impl std::error::Error> {
     use simple_logger::SimpleLogger;
     use winit::{
         event::{Event, WindowEvent},
-        event_loop::{ControlFlow, EventLoop},
+        event_loop::EventLoop,
         window::WindowBuilder,
     };
 
@@ -35,8 +35,6 @@ fn main() -> Result<(), impl std::error::Error> {
 
     event_loop.run(move |event, elwt| {
         println!("{event:?}");
-
-        elwt.set_control_flow(ControlFlow::Wait);
 
         match event {
             Event::WindowEvent {